Fix center crash (#4725)

+++ include/list_menu.h
@@ -40,9 +40,9 @@ struct ListMenuTemplate
     const struct ListMenuItem *items;
     void (* moveCursorFunc)(s32 itemIndex, bool8 onInit, struct ListMenu *list);
     void (* itemPrintFunc)(u8 windowId, u32 itemId, u8 y);
-    u16 totalItems:12;
-    u16 maxShowed:12;
-    u16 textNarrowWidth:8;
+    u32 totalItems:12;
+    u32 maxShowed:12;
+    u32 textNarrowWidth:8;
     u8 windowId;
     u8 header_X;
     u8 item_X;
